Refactor DB connector
