+++ src/datasource-zabbix/zabbix/connectors/sql/sqlConnector.ts
@@ -11,6 +11,9 @@ const supportedDatabases = {
 };
 
 export class SQLConnector extends DBConnector {
+  private limit: number;
+  private sqlDialect: any;
+
   constructor(options) {
     super(options);
 
@@ -35,28 +38,18 @@ export class SQLConnector extends DBConnector {
    * Try to invoke test query for one of Zabbix database tables.
    */
   testDataSource() {
-    let testQuery = this.sqlDialect.testQuery();
+    const testQuery = this.sqlDialect.testQuery();
     return this.invokeSQLQuery(testQuery);
   }
 
   getHistory(items, timeFrom, timeTill, options) {
-    let {intervalMs, consolidateBy} = options;
-    let intervalSec = Math.ceil(intervalMs / 1000);
-
-    // The interval must match the time range exactly n times, otherwise
-    // the resulting first and last data points will yield invalid values in the
-    // calculated average value in downsampleSeries - when using consolidateBy(avg)
-    let numOfIntervals = Math.ceil((timeTill - timeFrom) / intervalSec);
-    intervalSec = (timeTill - timeFrom) / numOfIntervals;
-
-    consolidateBy = consolidateBy || 'avg';
-    let aggFunction = dbConnector.consolidateByFunc[consolidateBy];
+    const { aggFunction, intervalSec } = getAggFunc(timeFrom, timeTill, options);
 
     // Group items by value type and perform request for each value type
-    let grouped_items = _.groupBy(items, 'value_type');
-    let promises = _.map(grouped_items, (items, value_type) => {
-      let itemids = _.map(items, 'itemid').join(', ');
-      let table = HISTORY_TO_TABLE_MAP[value_type];
+    const grouped_items = _.groupBy(items, 'value_type');
+    const promises = _.map(grouped_items, (items, value_type) => {
+      const itemids = _.map(items, 'itemid').join(', ');
+      const table = HISTORY_TO_TABLE_MAP[value_type];
       let query = this.sqlDialect.historyQuery(itemids, table, timeFrom, timeTill, intervalSec, aggFunction);
 
       query = compactQuery(query);
@@ -69,23 +62,14 @@ export class SQLConnector extends DBConnector {
   }
 
   getTrends(items, timeFrom, timeTill, options) {
-    let { intervalMs, consolidateBy } = options;
-    let intervalSec = Math.ceil(intervalMs / 1000);
-
-    // The interval must match the time range exactly n times, otherwise
-    // the resulting first and last data points will yield invalid values in the
-    // calculated average value in downsampleSeries - when using consolidateBy(avg)
-    let numOfIntervals = Math.ceil((timeTill - timeFrom) / intervalSec);
-    intervalSec = (timeTill - timeFrom) / numOfIntervals;
-
-    consolidateBy = consolidateBy || 'avg';
-    let aggFunction = dbConnector.consolidateByFunc[consolidateBy];
+    const { consolidateBy } = options;
+    const { aggFunction, intervalSec } = getAggFunc(timeFrom, timeTill, options);
 
     // Group items by value type and perform request for each value type
-    let grouped_items = _.groupBy(items, 'value_type');
-    let promises = _.map(grouped_items, (items, value_type) => {
-      let itemids = _.map(items, 'itemid').join(', ');
-      let table = TREND_TO_TABLE_MAP[value_type];
+    const grouped_items = _.groupBy(items, 'value_type');
+    const promises = _.map(grouped_items, (items, value_type) => {
+      const itemids = _.map(items, 'itemid').join(', ');
+      const table = TREND_TO_TABLE_MAP[value_type];
       let valueColumn = _.includes(['avg', 'min', 'max', 'sum'], consolidateBy) ? consolidateBy : 'avg';
       valueColumn = dbConnector.consolidateByTrendColumns[valueColumn];
       let query = this.sqlDialect.trendsQuery(itemids, table, timeFrom, timeTill, intervalSec, aggFunction, valueColumn);
@@ -100,7 +84,7 @@ export class SQLConnector extends DBConnector {
   }
 
   invokeSQLQuery(query) {
-    let queryDef = {
+    const queryDef = {
       refId: 'A',
       format: 'time_series',
       datasourceId: this.datasourceId,
@@ -116,7 +100,7 @@ export class SQLConnector extends DBConnector {
       }
     })
     .then(response => {
-      let results = response.data.results;
+      const results = response.data.results;
       if (results['A']) {
         return results['A'].frames;
       } else {
@@ -124,4 +108,20 @@ export class SQLConnector extends DBConnector {
       }
     });
   }
+}
+
+function getAggFunc(timeFrom, timeTill, options) {
+  const { intervalMs } = options;
+  let { consolidateBy } = options;
+  let intervalSec = Math.ceil(intervalMs / 1000);
+
+  // The interval must match the time range exactly n times, otherwise
+  // the resulting first and last data points will yield invalid values in the
+  // calculated average value in downsampleSeries - when using consolidateBy(avg)
+  const numOfIntervals = Math.ceil((timeTill - timeFrom) / intervalSec);
+  intervalSec = (timeTill - timeFrom) / numOfIntervals;
+
+  consolidateBy = consolidateBy || 'avg';
+  const aggFunction = dbConnector.consolidateByFunc[consolidateBy];
+  return { aggFunction, intervalSec };
 }
